refactor(extension/podman): mv LinuxSocketCompatibility to dedicated file
