+++ extensions/podman/packages/extension/src/compatibility-mode/linux-socket-compatibility.spec.ts
@@ -1,0 +1,53 @@
+/*********************************************************************
+ * Copyright (C) 2026 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************/
+import * as extensionApi from '@podman-desktop/api';
+import { beforeEach, expect, test, vi } from 'vitest';
+
+import { getSocketCompatibility } from '/@/compatibility-mode/compatibility-mode';
+import { LinuxSocketCompatibility } from '/@/compatibility-mode/linux-socket-compatibility';
+
+beforeEach(() => {
+  vi.resetAllMocks();
+
+  Object.defineProperty(process, 'platform', {
+    value: 'linux',
+  });
+});
+
+// Linux tests
+test('linux: compatibility mode pass', async () => {
+  expect(() => getSocketCompatibility()).toBeTruthy();
+});
+
+// Fail when trying to use runSystemdCommand with a command that's not "enable" or "disable"
+test('linux: fail when trying to use runSystemdCommand with a command that is not "enable" or "disable"', async () => {
+  const socketCompatClass = new LinuxSocketCompatibility();
+  await expect(socketCompatClass.runSystemdCommand('start', 'enabled')).rejects.toThrowError(
+    'runSystemdCommand only accepts enable or disable as the command',
+  );
+});
+
+test('linux: pass enabling when systemctl command exists', async () => {
+  vi.mocked(extensionApi.process.exec).mockResolvedValue({} as extensionApi.RunResult);
+
+  const socketCompatClass = new LinuxSocketCompatibility();
+
+  // Expect enable() to pass since systemctl command exists
+  await expect(socketCompatClass.enable()).resolves.toBeUndefined();
+  expect(extensionApi.window.showErrorMessage).not.toHaveBeenCalled();
+});
